Refactoring of the messy code
