+++ PortaudioRecording.h
@@ -1,0 +1,5 @@
+SAMPLE *retrieveResults();
+void startRecording();
+void initializePA();
+void resetPA();
+void cleanupPA();
